test(vibration): split VibrationSequencer tests into dedicated suite; add reusable MockVibrationOut; keep AlarmFlashScheduler suite focused
