+++ test/pure/test_alarm_flash_scheduler_pure/test_main.cpp
@@ -1,7 +1,5 @@
 #include <unity.h>
 #include "AlarmFlashScheduler.h"
-#include "VibrationSequencer.h"
-#include "IVibration.h"
 
 void setUp(void) {}
 void tearDown(void) {}
@@ -52,48 +50,6 @@ int main(int, char**) {
     RUN_TEST(test_750ms_is_off);
     RUN_TEST(test_4000ms_is_finished);
     RUN_TEST(test_after_finished_is_always_off);
-    // --- VibrationSequencer tests ---
-    {
-        // local mock
-        class MockVibrationOut : public IVibrationOutput {
-        public:
-            uint8_t lastDuty = 0;
-            void setDutyPercent(uint8_t dutyPercent) override { lastDuty = dutyPercent; }
-        };
-
-        auto test_starts_active = [](){
-            VibrationSequencer seq;
-            seq.loadPattern({ {100, 50}, {100, 0}, {100, 50} }, false);
-            seq.start(0);
-            TEST_ASSERT_TRUE(seq.isActive());
-        };
-        auto test_duty_in_first_segment = [](){
-            VibrationSequencer seq; MockVibrationOut out;
-            seq.loadPattern({ {100, 50}, {100, 0}, {100, 50} }, false);
-            seq.start(0);
-            seq.update(50, &out);
-            TEST_ASSERT_EQUAL_UINT8(50, out.lastDuty);
-        };
-        auto test_duty_in_second_segment = [](){
-            VibrationSequencer seq; MockVibrationOut out;
-            seq.loadPattern({ {100, 50}, {100, 0}, {100, 50} }, false);
-            seq.start(0);
-            seq.update(150, &out);
-            TEST_ASSERT_EQUAL_UINT8(0, out.lastDuty);
-        };
-        auto test_stops_after_oneshot = [](){
-            VibrationSequencer seq; MockVibrationOut out;
-            seq.loadPattern({ {100, 50}, {100, 0}, {100, 50} }, false);
-            seq.start(0);
-            seq.update(350, &out);
-            TEST_ASSERT_FALSE(seq.isActive());
-        };
-
-        RUN_TEST(test_starts_active);
-        RUN_TEST(test_duty_in_first_segment);
-        RUN_TEST(test_duty_in_second_segment);
-        RUN_TEST(test_stops_after_oneshot);
-    }
     return UNITY_END();
 }
